requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
black==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudinary
cryptography==38.0.4
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.51.0
django-crispy-forms
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
idna==3.4
ipython==8.7.0
jedi==0.18.2
jwt==1.3.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
oauthlib==3.2.2
parso==0.8.3
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.8.0
urllib3==1.26.13
wcwidth==0.2.5
